plugin_data() - Explicitly disable translations. WordPress 6.7.0
Function _load_textdomain_just_in_time was called incorrectly. Translation loading for the plugin domain was triggered too early. This is usually an indicator for some code in the plugin or theme running too early. Translations should be loaded at the init action or later. (This message was added in version 6.7.0.)